third_party/blink/web_tests/fast/dom/shadow/offsetWidth-host-style-change-expected.txt

offsetWidth of a fixed width element should cause a style recalc if host styles are invalid

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS div.offsetWidth is 100
PASS div.offsetWidth is 0
